+++ HISTORY.rst
@@ -15,6 +15,9 @@ Unreleased Changes
   will be logged when invoked on floating-point rasters, as using this on
   continuous rasters can result in excessive memory consumption. To use this
   feature, set ``include_value_counts=True`` when calling ``zonal_statistics``.
+* ``pygeoprocessing.get_gis_type`` will now raise a ``ValueError`` if the file
+  cannot be opened as ``gdal.OF_RASTER`` or ``gdal.OF_VECTOR``.
+  https://github.com/natcap/pygeoprocessing/issues/244
 * Fixing an error message in ``convolve_2d`` when the signal or kernel is a
   row based blocksize. https://github.com/natcap/pygeoprocessing/issues/228.